feat(tests): consolidate sorting tests into a single class and enhance test coverage
